backend/requirements.txt
anyio==3.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
cryptography==38.0.1
Deprecated==1.2.13
dill==0.3.6
Flask==2.2.2
Flask-Cors==3.0.10
gcloud==0.18.3
googleapis-common-protos==1.56.4
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauth2client==4.1.3
packaging==21.3
pipeline-ai==0.1.3
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pyhumps==3.8.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.21.0
python-jwt==4.0.0
replicate==0.4.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
